+++ frontend/src/pages/News/AllPosts.tsx
@@ -1,4 +1,3 @@
-import { Skeleton } from '@mui/material'
 import { useEffect, useState } from 'react'
 import { useUrlSearchParams, LinkWithStickyParams } from '../../utils/urlutils'
 import {
@@ -17,81 +16,10 @@ import { useQuery } from 'react-query'
 import { type Article } from './NewsPage'
 import SignupSection from '../ui/SignupSection'
 import { Link } from 'react-router-dom'
+import HetPostsLoading from '../../styles/HetComponents/HetPostsLoading'
+import PinnedArticles from './PinnedArticles'
 
 export const ARTICLES_TERM = 'Articles'
-const NUM_OF_LOADING_SKELETONS = 12
-
-/*
-displays several loading indicator elements while blog content is fetched
-*/
-
-export function ArticlesSkeleton(props: {
-  doPulse: boolean
-  numberLoading?: number
-}) {
-  const numberLoadingSkeletons = props.numberLoading ?? NUM_OF_LOADING_SKELETONS
-
-  return (
-    <div className='flex flex-wrap justify-between'>
-      {[...Array(numberLoadingSkeletons)].map((_, i) => {
-        return (
-          <div
-            className='flex w-full flex-col items-center sm:w-1/2 xl:w-1/3'
-            key={i}
-          >
-            <Skeleton
-              animation={props.doPulse && 'wave'}
-              variant='rectangular'
-              height={100}
-              width={150}
-            ></Skeleton>
-            <Skeleton
-              animation={false}
-              variant='text'
-              height={36}
-              width={200}
-            ></Skeleton>
-            <div className='mb-10'>
-              <Skeleton
-                animation={props.doPulse && 'wave'}
-                variant='text'
-                height={36}
-                width={175}
-              ></Skeleton>
-            </div>
-          </div>
-        )
-      })}
-    </div>
-  )
-}
-
-interface PinnedArticlesProps {
-  articles: Article[]
-}
-
-function PinnedArticles(props: PinnedArticlesProps) {
-  const { articles } = props
-
-  return articles?.length > 0 ? (
-    <div className='shadow-md'>
-      <h6 className='m-0 text-center font-serif font-light text-alt-green'>
-        Featured:
-      </h6>
-      <div className='flex'>
-        {articles.map((post: any) => {
-          return (
-            <div className='w-full sm:w-1/2' key={post.id}>
-              <NewsPreviewCard article={post} />
-            </div>
-          )
-        })}
-      </div>
-    </div>
-  ) : (
-    <></>
-  )
-}
 
 function AllPosts() {
   // articles matching client applied filters (author, category, etc)
@@ -359,22 +287,25 @@ function AllPosts() {
               })}
             </div>
 
-            <div className='flex flex-col justify-center'>
+            <div className='flex flex-wrap items-start justify-between'>
               {isLoading && (
                 <>
-                  <ArticlesSkeleton doPulse={true} />
+                  <HetPostsLoading
+                    doPulse={true}
+                    className='w-full sm:w-1/2 lg:w-1/3 xl:w-1/4'
+                  />
                   <div className='m-10'>
                     <i>Updating articles...</i>
                   </div>
                 </>
               )}
               {error && !isLoading && (
-                <>
+                <div className='w-full sm:w-1/2 lg:w-1/3 xl:w-1/4'>
                   <div className='m-10'>
                     <i>Problem updating articles.</i>
                   </div>
-                  <ArticlesSkeleton doPulse={false} />
-                </>
+                  <HetPostsLoading doPulse={false} />
+                </div>
               )}
             </div>
           </div>
